third_party/WebKit/LayoutTests/fast/dom/Range/deleteContents-doctype.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="container">
</div>
<script>
description("This test verifies if Range.deleteContents can handle doctype.");

var range = document.createRange();
range.setStart(document, 0);
range.setEnd(document, 1);

shouldNotThrow("range.deleteContents()");
</script>
